made the Classes and Program.cs to be more generic so that all the projects can use the same files and each class can import the class they need
