+++ KIT206/KIT206.csproj
@@ -5,4 +5,13 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Compile Include="..\Solution Files\Class.cs" Link="Class.cs" />
+    <Compile Include="..\Solution Files\enums.cs" Link="enums.cs" />
+    <Compile Include="..\Solution Files\Meeting.cs" Link="Meeting.cs" />
+    <Compile Include="..\Solution Files\Program.cs" Link="Program.cs" />
+    <Compile Include="..\Solution Files\Student.cs" Link="Student.cs" />
+    <Compile Include="..\Solution Files\StudentGroup.cs" Link="StudentGroup.cs" />
+  </ItemGroup>
+
 </Project>
